requirements.txt
streamlit
pandas
numpy
scikit-learn
pillow
matplotlib
seaborn
plotly
